Connect modules using CommonJS dependencies main.js
