+++ src/index.html
@@ -128,8 +128,6 @@
   </a>
 </template>
 
-<script src="js/resizer.js" charset="utf-8"></script>
-<script src="js/upload.js" charset="utf-8"></script>
-<script src="js/pictures.js" charset="utf-8"></script>
+<script src="js/main.js" charset="utf-8"></script>
 </body>
 </html>
